refactor: удали прямое использование LS из компонентов
