+++ src/components/SearchHistory.jsx
@@ -1,48 +1,28 @@
 /* eslint-disable no-console */
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { Button } from "@mui/material";
 import IconButton from "@mui/material/IconButton";
 import DeleteIcon from "@mui/icons-material/Delete";
 import styles from "./SearchHistory.module.css";
+import { useSearchHistoryManagement } from "../hooks/useSearchHistoryManagement";
 
 export function SearchHistory() {
   const userHistory = useSelector((state) => state.auth.email) || "";
   const navigate = useNavigate();
-  const [historyUpdated, setHistoryUpdated] = useState(false);
+  const { getSearchHistory, clearHistory, clearHistoryItem } =
+    useSearchHistoryManagement(userHistory);
 
   const handleHistoryClick = (search) => {
     navigate(`/search?name=${search}`);
   };
 
-  const moviesHistory = JSON.parse(localStorage.getItem("moviesHistory")) || {};
+  const moviesHistory = getSearchHistory();
   const searchArray = moviesHistory[userHistory] || [];
-
-  const clearHistory = () => {
-    const updatedMoviesHistory = { ...moviesHistory };
-    delete updatedMoviesHistory[userHistory];
-    localStorage.setItem("moviesHistory", JSON.stringify(updatedMoviesHistory));
-    setHistoryUpdated(true);
-  };
-
-  useEffect(() => {}, [historyUpdated]);
 
   const keysOfSearches = searchArray.map((item) => Object.keys(item)) || [];
   const userSearches = searchArray.map((item) => Object.values(item)) || [];
-
-  const clearHistoryItem = (search) => {
-    const updatedMoviesHistory = { ...moviesHistory };
-
-    const updatedUserSearches = searchArray.filter(
-      (item) => Object.keys(item)[0].toString() !== search.toString(),
-    );
-    updatedMoviesHistory[userHistory] = updatedUserSearches;
-
-    localStorage.setItem("moviesHistory", JSON.stringify(updatedMoviesHistory));
-    setHistoryUpdated(!historyUpdated);
-    console.log(`вы удалили ${search}`);
-  };
 
   if (searchArray.length > 0 && userSearches.length > 0) {
     return (
