+++ src/utilities/getFavoriteMoviesData.js
@@ -1,0 +1,2 @@
+export const getFavoriteMoviesData = () =>
+  JSON.parse(localStorage.getItem("favoriteMoviesId")) || {};
