+++ src/utilities/getStoredIsFavorite.js
@@ -1,0 +1,2 @@
+export const getStoredIsFavorite = (value) =>
+  localStorage.getItem(`isFavorite_${value}`);
